feat(query-orchestrator): introduce unload without temp table (#5324)

+++ packages/cubejs-query-orchestrator/src/orchestrator/PreAggregations.ts
@@ -18,7 +18,8 @@ import { cancelCombinator, SaveCancelFn, DriverInterface, BaseDriver,
   DownloadTableData,
   InlineTable,
   StreamOptions,
-  UnloadOptions } from '@cubejs-backend/base-driver';
+  UnloadOptions,
+  DriverCapabilities } from '@cubejs-backend/base-driver';
 import { RedisCacheDriver } from './RedisCacheDriver';
 import { LocalCacheDriver } from './LocalCacheDriver';
 import { Query, QueryCache, QueryTuple, QueryWithParams } from './QueryCache';
@@ -104,6 +105,20 @@ type VersionEntry = {
   build_range_end?: string,
   // eslint-disable-next-line camelcase
   naming_version?: number
+};
+
+type IndexesSql = { sql: [string, unknown[]], indexName: string }[];
+type InvalidationKeys = unknown[];
+
+type QueryKey = [QueryTuple, IndexesSql, InvalidationKeys] | [QueryTuple, InvalidationKeys];
+
+type QueryOptions = {
+  queryKey: QueryKey;
+  newVersionEntry: VersionEntry;
+  query: string;
+  values: unknown[];
+  requestId: string;
+  buildRangeEnd?: string;
 };
 
 type TableCacheEntry = {
@@ -747,26 +762,30 @@ export class PreAggregationLoader {
     );
   }
 
-  protected preAggregationQueryKey(invalidationKeys) {
+  protected preAggregationQueryKey(invalidationKeys: InvalidationKeys): QueryKey {
     return this.preAggregation.indexesSql && this.preAggregation.indexesSql.length ?
       [this.preAggregation.loadSql, this.preAggregation.indexesSql, invalidationKeys] :
       [this.preAggregation.loadSql, invalidationKeys];
   }
 
-  protected targetTableName(versionEntry): string {
+  protected targetTableName(versionEntry: VersionEntry): string {
     // eslint-disable-next-line no-use-before-define
     return PreAggregations.targetTableName(versionEntry);
   }
 
-  public refresh(newVersionEntry, invalidationKeys, client) {
-    let refreshStrategy = this.refreshImplStoreInSourceStrategy;
+  public refresh(newVersionEntry: VersionEntry, invalidationKeys: InvalidationKeys, client) {
+    let refreshStrategy = this.refreshStoreInSourceStrategy;
     if (this.preAggregation.external) {
       const readOnly =
         this.preAggregation.readOnly ||
         client.config && client.config.readOnly ||
         client.readOnly && (typeof client.readOnly === 'boolean' ? client.readOnly : client.readOnly());
-      refreshStrategy = readOnly ?
-        this.refreshImplStreamExternalStrategy : this.refreshImplTempTableExternalStrategy;
+
+      if (readOnly) {
+        refreshStrategy = this.refreshReadOnlyExternalStrategy;
+      } else {
+        refreshStrategy = this.refreshWriteStrategy;
+      }
     }
     return cancelCombinator(
       saveCancelFn => refreshStrategy.bind(this)(
@@ -785,7 +804,7 @@ export class PreAggregationLoader {
     );
   }
 
-  protected queryOptions(invalidationKeys, query, params, targetTableName, newVersionEntry) {
+  protected queryOptions(invalidationKeys: InvalidationKeys, query: string, params: unknown[], targetTableName: string, newVersionEntry: VersionEntry) {
     return {
       queryKey: this.preAggregationQueryKey(invalidationKeys),
       query,
@@ -797,11 +816,11 @@ export class PreAggregationLoader {
     };
   }
 
-  protected async refreshImplStoreInSourceStrategy(
+  protected async refreshStoreInSourceStrategy(
     client: DriverInterface,
-    newVersionEntry,
+    newVersionEntry: VersionEntry,
     saveCancelFn: SaveCancelFn,
-    invalidationKeys
+    invalidationKeys: InvalidationKeys
   ) {
     const [loadSql, params] =
         Array.isArray(this.preAggregation.loadSql) ? this.preAggregation.loadSql : [this.preAggregation.loadSql, []];
@@ -832,39 +851,53 @@ export class PreAggregationLoader {
     }
   }
 
-  /**
-   * Strategy to copy pre-aggregation from source db (with write permissions) to external data
-   */
-  protected async refreshImplTempTableExternalStrategy(
+  protected async refreshWriteStrategy(
     client: DriverInterface,
     newVersionEntry: VersionEntry,
     saveCancelFn: SaveCancelFn,
-    invalidationKeys
+    invalidationKeys: InvalidationKeys,
   ) {
-    const [loadSql, params] =
-        Array.isArray(this.preAggregation.loadSql) ? this.preAggregation.loadSql : [this.preAggregation.loadSql, []];
+    const capabilities = client?.capabilities();
+
+    const withTempTable = !(capabilities?.unloadWithoutTempTable);
+
+    return this.runWriteStrategy(
+      client,
+      newVersionEntry,
+      saveCancelFn,
+      invalidationKeys,
+      withTempTable,
+    );
+  }
+
+  /**
+   * Runs export strategy with write access in data source
+   */
+  protected async runWriteStrategy(
+    client: DriverInterface,
+    newVersionEntry: VersionEntry,
+    saveCancelFn: SaveCancelFn,
+    invalidationKeys: InvalidationKeys,
+    withTempTable: boolean
+  ) {
     await client.createSchemaIfNotExists(this.preAggregation.preAggregationsSchema);
     const targetTableName = this.targetTableName(newVersionEntry);
-    const query = QueryCache.replacePreAggregationTableNames(loadSql, this.preAggregationsTablesToTempTables)
-      .replace(
-        this.preAggregation.tableName,
-        targetTableName
-      );
-    const queryOptions = this.queryOptions(invalidationKeys, query, params, targetTableName, newVersionEntry);
-    this.logExecutingSql(queryOptions);
-    await saveCancelFn(client.loadPreAggregationIntoTable(
+    const queryOptions = await this.prepareWriteStrategy(
+      client,
       targetTableName,
-      query,
-      params,
-      queryOptions
-    ));
+      newVersionEntry,
+      saveCancelFn,
+      invalidationKeys,
+      withTempTable,
+    );
 
     try {
-      const tableData = await this.downloadTempExternalPreAggregation(
+      const tableData = await this.downloadExternalPreAggregation(
         client,
         newVersionEntry,
         saveCancelFn,
-        queryOptions
+        queryOptions,
+        withTempTable,
       );
 
       try {
@@ -875,25 +908,86 @@ export class PreAggregationLoader {
         }
       }
     } finally {
+      await this.cleanupWriteStrategy(
+        client,
+        targetTableName,
+        queryOptions,
+        saveCancelFn,
+        withTempTable
+      );
+    }
+  }
+
+  /**
+   * Cleanup tables after write strategy
+   */
+  protected async cleanupWriteStrategy(
+    client: DriverInterface,
+    targetTableName: string,
+    queryOptions: QueryOptions,
+    saveCancelFn: SaveCancelFn,
+    withTempTable: boolean
+  ) {
+    if (withTempTable) {
       const actualTables = await client.getTablesQuery(this.preAggregation.preAggregationsSchema);
       const mappedActualTables = actualTables.map(t => `${this.preAggregation.preAggregationsSchema}.${t.table_name || t.TABLE_NAME}`);
       if (mappedActualTables.includes(targetTableName)) {
         await client.dropTable(targetTableName);
       }
-      // We must clean orphaned in any cases: success or exception
-      await this.loadCache.fetchTables(this.preAggregation);
-      await this.dropOrphanedTables(client, targetTableName, saveCancelFn, false, queryOptions);
+    }
+
+    // We must clean orphaned in any cases: success or exception
+    await this.loadCache.fetchTables(this.preAggregation);
+    await this.dropOrphanedTables(client, targetTableName, saveCancelFn, false, queryOptions);
+  }
+
+  /**
+   * Create table (if required) and prepares query options object
+   */
+  protected async prepareWriteStrategy(
+    client: DriverInterface,
+    targetTableName: string,
+    newVersionEntry: VersionEntry,
+    saveCancelFn: SaveCancelFn,
+    invalidationKeys: InvalidationKeys,
+    withTempTable: boolean
+  ): Promise<QueryOptions> {
+    if (withTempTable) {
+      const [loadSql, params] =
+      Array.isArray(this.preAggregation.loadSql) ? this.preAggregation.loadSql : [this.preAggregation.loadSql, []];
+  
+      const query = QueryCache.replacePreAggregationTableNames(loadSql, this.preAggregationsTablesToTempTables)
+        .replace(
+          this.preAggregation.tableName,
+          targetTableName
+        );
+      const queryOptions = this.queryOptions(invalidationKeys, query, params, targetTableName, newVersionEntry);
+      this.logExecutingSql(queryOptions);
+      await saveCancelFn(client.loadPreAggregationIntoTable(
+        targetTableName,
+        query,
+        params,
+        queryOptions
+      ));
+
+      return queryOptions;
+    } else {
+      const [sql, params] =
+        Array.isArray(this.preAggregation.sql) ? this.preAggregation.sql : [this.preAggregation.sql, []];
+      const queryOptions = this.queryOptions(invalidationKeys, sql, params, targetTableName, newVersionEntry);
+      this.logExecutingSql(queryOptions);
+      return queryOptions;
     }
   }
 
   /**
    * Strategy to copy pre-aggregation from source db (for read-only permissions) to external data
    */
-  protected async refreshImplStreamExternalStrategy(
+  protected async refreshReadOnlyExternalStrategy(
     client: DriverInterface,
     newVersionEntry: VersionEntry,
-    saveCancelFn,
-    invalidationKeys
+    saveCancelFn: SaveCancelFn,
+    invalidationKeys: InvalidationKeys
   ) {
     const [sql, params] =
         Array.isArray(this.preAggregation.sql) ? this.preAggregation.sql : [this.preAggregation.sql, []];
@@ -948,19 +1042,15 @@ export class PreAggregationLoader {
   }
 
   /**
-   * Create table (for db with write permissions) and extract data via memory/stream/unload
+   * prepares download data for future cube store usage
    */
-  protected async downloadTempExternalPreAggregation(
+  protected async downloadExternalPreAggregation(
     client: DriverInterface,
-    newVersionEntry,
+    newVersionEntry: VersionEntry,
     saveCancelFn: SaveCancelFn,
-    queryOptions: any
+    queryOptions: QueryOptions,
+    withTempTable: boolean
   ) {
-    // @todo Deprecated, BaseDriver already implements it, before remove we need to add check for factoryDriver
-    if (!client.downloadTable) {
-      throw new Error('Can\'t load external pre-aggregation: source driver doesn\'t support downloadTable()');
-    }
-
     const table = this.targetTableName(newVersionEntry);
     this.logger('Downloading external pre-aggregation', queryOptions);
 
@@ -969,31 +1059,12 @@ export class PreAggregationLoader {
       const capabilities = externalDriver.capabilities && externalDriver.capabilities();
 
       let tableData: DownloadTableData;
-
-      if (capabilities.csvImport && client.unload && await client.isUnloadSupported(this.getUnloadOptions())) {
-        tableData = await saveCancelFn(client.unload(table, this.getUnloadOptions()));
-      } else if (capabilities.streamImport && client.stream) {
-        tableData = await saveCancelFn(
-          client.stream(`SELECT * FROM ${table}`, [], this.getStreamingOptions())
-        );
-
-        if (client.unload) {
-          const stream = new LargeStreamWarning(this.preAggregation.preAggregationId, (msg) => {
-            this.logger('Downloading external pre-aggregation warning', {
-              ...queryOptions,
-              error: msg
-            });
-          });
-          tableData.rowStream.pipe(stream);
-          tableData.rowStream = stream;
-        }
+      if (withTempTable) {
+        tableData = await this.getTableDataWithTempTable(client, table, saveCancelFn, queryOptions, capabilities);
       } else {
-        tableData = await saveCancelFn(client.downloadTable(table, capabilities));
+        tableData = await this.getTableDataWithoutTempTable(client, table, saveCancelFn, queryOptions, capabilities);
       }
 
-      if (!tableData.types) {
-        tableData.types = await saveCancelFn(client.tableColumnTypes(table));
-      }
       this.logger('Downloading external pre-aggregation completed', queryOptions);
 
       return tableData;
@@ -1006,11 +1077,77 @@ export class PreAggregationLoader {
     }
   }
 
+  /**
+   * prepares download data when temp table = true
+   */
+  protected async getTableDataWithTempTable(client: DriverInterface, table: string, saveCancelFn: SaveCancelFn, queryOptions: QueryOptions, externalDriverCapabilities: DriverCapabilities) {
+    let tableData: DownloadTableData;
+
+    if (externalDriverCapabilities.csvImport && client.unload && await client.isUnloadSupported(this.getUnloadOptions())) {
+      tableData = await saveCancelFn(client.unload(table, this.getUnloadOptions()));
+    } else if (externalDriverCapabilities.streamImport && client.stream) {
+      tableData = await saveCancelFn(
+        client.stream(`SELECT * FROM ${table}`, [], this.getStreamingOptions())
+      );
+
+      if (client.unload) {
+        const stream = new LargeStreamWarning(this.preAggregation.preAggregationId, (msg) => {
+          this.logger('Downloading external pre-aggregation warning', {
+            ...queryOptions,
+            error: msg
+          });
+        });
+        tableData.rowStream.pipe(stream);
+        tableData.rowStream = stream;
+      }
+    } else {
+      tableData = await saveCancelFn(client.downloadTable(table, externalDriverCapabilities));
+    }
+
+    if (!tableData.types) {
+      tableData.types = await saveCancelFn(client.tableColumnTypes(table));
+    }
+
+    return tableData;
+  }
+
+  /**
+   * prepares download data when temp table = false
+   */
+  protected async getTableDataWithoutTempTable(client: DriverInterface, table: string, saveCancelFn: SaveCancelFn, queryOptions: QueryOptions, externalDriverCapabilities: DriverCapabilities) {
+    const [sql, params] =
+        Array.isArray(this.preAggregation.sql) ? this.preAggregation.sql : [this.preAggregation.sql, []];
+
+    let tableData: DownloadTableData;
+    if (externalDriverCapabilities.csvImport && client.unload && await client.isUnloadSupported(this.getUnloadOptions())) {
+      return saveCancelFn(client.unload(table, { ...this.getUnloadOptions(), query: { sql, params } }));
+    } else if (externalDriverCapabilities.streamImport && client.stream) {
+      tableData = await saveCancelFn(
+        client.stream(sql, params, this.getStreamingOptions())
+      );
+
+      if (client.unload) {
+        const stream = new LargeStreamWarning(this.preAggregation.preAggregationId, (msg) => {
+          this.logger('Downloading external pre-aggregation warning', {
+            ...queryOptions,
+            error: msg
+          });
+        });
+        tableData.rowStream.pipe(stream);
+        tableData.rowStream = stream;
+      }
+    } else {
+      tableData = { rows: await saveCancelFn(client.query(sql, params)) };
+    }
+
+    return tableData;
+  }
+
   protected async uploadExternalPreAggregation(
     tableData: DownloadTableData,
     newVersionEntry: VersionEntry,
     saveCancelFn: SaveCancelFn,
-    queryOptions: any
+    queryOptions: QueryOptions
   ) {
     const externalDriver: DriverInterface = await this.externalDriverFactory();
     const table = this.targetTableName(newVersionEntry);
@@ -1025,7 +1162,7 @@ export class PreAggregationLoader {
         this.preAggregation.uniqueKeyColumns,
         queryOptions,
         this.preAggregation.aggregationsColumns,
-        this.prepareCreateTableIndexes(newVersionEntry, queryOptions),
+        this.prepareCreateTableIndexes(newVersionEntry),
       )
     ).catch((error: any) => {
       this.logger('Uploading external pre-aggregation error', { ...queryOptions, error: error?.stack || error?.message });
@@ -1037,7 +1174,7 @@ export class PreAggregationLoader {
     await this.dropOrphanedTables(externalDriver, table, saveCancelFn, true, queryOptions);
   }
 
-  protected async createIndexes(driver, newVersionEntry, saveCancelFn, queryOptions) {
+  protected async createIndexes(driver, newVersionEntry: VersionEntry, saveCancelFn: SaveCancelFn, queryOptions: QueryOptions) {
     const indexesSql = this.prepareIndexesSql(newVersionEntry, queryOptions);
     for (let i = 0; i < indexesSql.length; i++) {
       const [query, params] = indexesSql[i].sql;
@@ -1045,7 +1182,7 @@ export class PreAggregationLoader {
     }
   }
 
-  protected prepareIndexesSql(newVersionEntry, queryOptions) {
+  protected prepareIndexesSql(newVersionEntry: VersionEntry, queryOptions: QueryOptions) {
     if (!this.preAggregation.indexesSql || !this.preAggregation.indexesSql.length) {
       return [];
     }
@@ -1067,7 +1204,7 @@ export class PreAggregationLoader {
     });
   }
 
-  protected prepareCreateTableIndexes(newVersionEntry, queryOptions) {
+  protected prepareCreateTableIndexes(newVersionEntry: VersionEntry) {
     if (!this.preAggregation.createTableIndexes || !this.preAggregation.createTableIndexes.length) {
       return [];
     }
@@ -1085,7 +1222,7 @@ export class PreAggregationLoader {
     justCreatedTable: string,
     saveCancelFn: SaveCancelFn,
     external: boolean,
-    queryOptions: any
+    queryOptions: QueryOptions
   ) {
     await this.preAggregations.addTableUsed(justCreatedTable);
